Move the GetVoteStats utils under the lib namespace
