+++ lib/cast_vote.js
@@ -4,7 +4,7 @@ const Shapes = require('./shapes')
 const { fetchProp } = require('./utils')
 const { ValidationError, NotFoundError } = require('./errors')
 const CastVoteService = require('../services/cast_vote')
-const GetVoteStats = require('../services/get_vote_stats')
+const GetVoteStats = require('./get_vote_stats')
 const SavePollRating = require('../services/save_poll_rating')
 const Reply = require('./reply')
