+++ lib/get_vote_stats.js
@@ -1,7 +1,7 @@
 const Assert = require('assert-plus')
-const Vote = require('../../lib/entities/sys/vote')
+const Vote = require('./entities/sys/vote')
 const groupBy = require('lodash.groupby')
-const { fetchProp, countMatching } = require('../../lib/utils')
+const { fetchProp, countMatching } = require('./utils')
 
 class GetVoteStats {
   static async forPoll(args, ctx, opts = {}) {
