+++ test/lib/get_vote_stats.js
